requirements.txt
python-dotenv
flask
pandas
python-dateutil
pymongo
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
apscheduler
dnspython
